refactor(heart): rename entity and package to heart from like
