+++ src/main/java/com/e2i/wemeet/domain/heart/HeartRepository.java
@@ -1,0 +1,7 @@
+package com.e2i.wemeet.domain.heart;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HeartRepository extends JpaRepository<Heart, Long> {
+
+}
