--- src/main/java/com/e2i/wemeet/domain/like/LikeRepository.java
@@ -1,7 +1,0 @@
-package com.e2i.wemeet.domain.like;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LikeRepository extends JpaRepository<Heart, Long> {
-
-}
